+++ src/containers/ModifyContract/Params.js
@@ -161,7 +161,7 @@ const ModifyParams = ({ nextStep, prevStep, handleStepChanges, params }) => {
       })
       .test('increasing', t('NO_POWER_INCREASING'), function () {
         return rates[this.parent.tariff]?.increasing
-          ? parseInt(this.parent.power2) >= parseInt(this.parent.power)
+          ? parseFloat(this.parent.power2) >= parseFloat(this.parent.power)
           : true
       })
       .test({
@@ -197,7 +197,7 @@ const ModifyParams = ({ nextStep, prevStep, handleStepChanges, params }) => {
       })
       .test('increasing', t('NO_POWER_INCREASING'), function () {
         return rates[this.parent.tariff]?.increasing
-          ? parseInt(this.parent.power3) >= parseInt(this.parent.power2)
+          ? parseFloat(this.parent.power3) >= parseFloat(this.parent.power2)
           : true
       })
       .test({
@@ -233,7 +233,7 @@ const ModifyParams = ({ nextStep, prevStep, handleStepChanges, params }) => {
       })
       .test('increasing', t('NO_POWER_INCREASING'), function () {
         return rates[this.parent.tariff]?.increasing
-          ? parseInt(this.parent.power4) >= parseInt(this.parent.power3)
+          ? parseFloat(this.parent.power4) >= parseFloat(this.parent.power3)
           : true
       })
       .test({
@@ -269,7 +269,7 @@ const ModifyParams = ({ nextStep, prevStep, handleStepChanges, params }) => {
       })
       .test('increasing', t('NO_POWER_INCREASING'), function () {
         return rates[this.parent.tariff]?.increasing
-          ? parseInt(this.parent.power5) >= parseInt(this.parent.power4)
+          ? parseFloat(this.parent.power5) >= parseFloat(this.parent.power4)
           : true
       })
       .test({
@@ -305,7 +305,7 @@ const ModifyParams = ({ nextStep, prevStep, handleStepChanges, params }) => {
       })
       .test('increasing', t('NO_POWER_INCREASING'), function () {
         return rates[this.parent.tariff]?.increasing
-          ? parseInt(this.parent.power6) >= parseInt(this.parent.power5)
+          ? parseFloat(this.parent.power6) >= parseFloat(this.parent.power5)
           : true
       })
       .test({
